Initialization and update

KitePodModels.init_kcu — Method

init_kcu(kcu::KCU, set::KiteUtils.Settings)

Inititalze the model of the kite control unit (KCU). The actual and the set values of depower are initialized to set.depower_offset * 0.01, the actual and the set values of steering to zero.

KitePodModels.on_timer — Function

on_timer(kcu::KCU, dt = 0.0)

Must be called at each clock tick. Parameter: Δt in seconds Updates the current values of steering and depower depending on the set values and the last value. If dt == 0.0, then it will be set to 1.0 / kcu.set.sample_freq .

Input and Output

KitePodModels.set_depower_steering — Method

set_depower_steering(kcu::KCU, depower, steering)

Set the values of depower and steering. The value for depower must be between 0.0 and 1.0, the value for steering between -1.0 and +1.0 .

KitePodModels.get_depower — Method

get_depower(kcu::KCU)

Read the current depower value. Result will be between 0.0 and 1.0.

KitePodModels.get_steering — Method

get_steering(kcu::KCU)

Read the current depower value. Result will be between -1.0 and 1.0.

Conversions

KitePodModels.calc_alpha_depower — Method

calc_alpha_depower(kcu::KCU, rel_depower)

Calculate the change of the angle between the kite and the last tether segment [rad] as function of the actual rel_depower value.

Returns nothing in case of error.