Initialization and update

KitePodSimulator.init_kcu — Method

init_kcu()

Inititalze the simulator. Must be called at the beginning of the simulation. The actual and the set values of depower are initialized to set.depower_offset * 0.01, the actual and the set values of steering to zero.

KitePodSimulator.on_timer — Function

on_timer(dt = 1.0 / set.sample_freq)

Must be called at each clock tick. Parameter: Δt in seconds Updates the current values of steering and depower depending on the set values and the last value.

Input and Output

KitePodSimulator.set_depower_steering — Method

set_depower_steering(depower, steering)

Set the values of depower and steering. The value for depower must be between 0.0 and 1.0, the value for steering between -1.0 and +1.0 .

KitePodSimulator.get_depower — Method

get_depower()

Read the current depower value. Result will be between 0.0 and 1.0.

KitePodSimulator.get_steering — Method

get_steering()

Read the current depower value. Result will be between -1.0 and 1.0.

Conversions

KitePodSimulator.calc_alpha_depower — Method

calc_alpha_depower(rel_depower)

Calculate the change of the angle between the kite and the last tether segment [rad] as function of the actual rel_depower value.

Returns nothing in case of error.